Extract parse helpers from CommandLineBuilder to OptionParsers
Moved ParseVerbosity, ParseTipLevel, ParseSectionList, and
ParseNuGetSourceOptions to a dedicated OptionParsers class. These are
reusable parsing utilities, not CLI wiring.

CommandLineBuilder retains thin forwarding methods for backward
compatibility with existing tests.
